feat(cli): naming refactor
